# Enable regex source gen / compiler vectorization of all sets